Redo the way jobs are submitted to HTcondor.
The generation of input files is now handled on the worker node, not locally.
This should speed up proceedings and cause less stress on the login node IO.
It also stops the user from having lots of input and tgz files. The only files
that should be returned are the spectrum files, zipped up.